// src/types/api.ts
export interface ModelCron {
    command: string
    /** 计划任务覆盖范围 (0:仅覆盖特定服务器 1:仅忽略特定服务器 2:由触发该计划任务的服务器执行) */
    cover: number
    created_at: string
    cron_job_id: number
    id: number
    /** 最后一次执行时间 */
    last_executed_at: string
    /** 最后一次执行结果 */
    last_result: boolean
    name: string
    /** 指定通知方式的分组 */
    notification_group_id: number
    /** 推送成功的通知 */
    push_successful: boolean
    /** 分钟 小时 天 月 星期 */
    scheduler: string
    servers: number[]
    /** 0:计划任务 1:触发任务 */
    task_type: number
    updated_at: string
}